depthchargectl: bless: Support partitions inside images
The other commands already support disk images where it makes sense, but
I had forgotten to implement it for bless. Having one positional
argument for two types of things makes things a bit complicated as
always, but it's managable. Adds a --partno argument for the partition
number with disk images, and -i for its short form as cgpt has that.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>