depthchargectl: build: Handle multiple roots from argument and kernel
Now that we have a root argument it should override the ones in the
kernel cmdline if explicitly given. We can't just append it to cmdline
because it might be the value we read from cmdline, resulting in
duplication. Instead, remove all other root= cmdline params, and then
append one if no existing one matches the root argument.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>