utils: argparse: Don't use ArgumentParser.error()
The usage message that it prints would have been fine if it's the first
message a command prints, but when it's in the middle of other messages
it's quite ugly. Also, we can't see the error traceback with it even in
debug loglevel. Looks like it's better to just not use it, and print
these errors the way others are printed.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>