depthchargectl: Allow board to be None at expense of some functionality
The build and check commands can obviously not work when no board is
given, but bless, list, remove, target should probably not require one
at all. I think I can get away with making the write command work only
with an unchecked image in this case.

This is mostly a first step to enable using depthchargectl commands on
non-ChromeOS hardware where we shouldn't be trying to manage the system
even if our commands are called e.g. by system hooks.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>