config.ini: Merge depthchargectl/build configs into depthchargectl
It's hard to override ignore-initramfs when it is set in the build
subsection, since it will default to False in the write subcommand and
get passed as False to the build subcommand, overriding the config
value. One has to override it twice, once in the depthchargectl section
so it's inherited to write, and once more in the depthchargectl/build
section so that it affects the build subcommand.

Instead let's just put the built-in value in the depthchargectl section
which means overriding at that one section will be enough, and move the
kernel-cmdline there as well for consistency.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>