depthchargectl: Match device-tree compatibles more precisely
The upstreamed device-tree compatibles for the qc7180 boards are a bit
weird. There's a long email thread [1] on the kernel mailing lists about
the reasoning behind how they are structured.

The gist of it is that the "google,<board>" compatible is no longer in
all the dtbs related to the board. We have to handle the "-rev*" and
"-sku*" suffixes, and hwid-match patterns we have can be for multiple
rev/sku combinations. Even worse, the dt compatible for "limozeen"
boards is prefixed with "google,lazor" instead of "google,limozeen".

Make it possible to specify a regex pattern as the dt-compatible board
config option. Replace the "dt-compatible = True" hack (which would
include all dtb files) with a catch-all regex pattern. If the config's
value doesn't look like a regex, implicitly add a segment to match all
rev and sku suffixes, so we can keep using "google,<board>".

[1] https://lore.kernel.org/all/CAD=FV=Vx5g_xTRZGc9wW=ZLnfsOcubTYFcnYQRC5jLm+n3en0w@mail.gmail.com/T/#u

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>