depthchargectl: Match board codename argument more fuzzily
The "--board" argument has so far matched the codenames exactly. This
causes a small annoyance because some old boards are not in the usual
codename format, like "veyron_minnie" instead of "minnie", or "peach-pi"
instead of "pi".

To make it possible to replace the longer forms with the shorter forms
in the board database, implement a fuzzy matching mechanism for the
codename argument based on parts split by dashes or underscores. This
lets us use e.g. both "veyron_minnie" and "minnie" for the same board,
or "rk3399-gru-kevin" and "gru_kevin" and "kevin" for the same board.
One unwanted quirk with it is "li" being understood as "falco_li".

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>