depthchargectl: build: Don't set kernel defaults with custom kernel
The build subcommand supports using the installed kernel parts, or
specifying custom kernel parts. The custom kernel arguments override the
parts from installed kernel, but the parts specified still default to
the installation. This leads to using the installed amd64 initramfs
while building for an arm64 board without an initramfs argument.

When a custom kernel is specified, we cannot know if the defaults from
the installed system will be usable with that. Don't set the defaults
in that case and expect the other arguments to be passed as well.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>