depthchargectl: list: Don't fail early when a disk raises cpgt error
When calling the list subcommand with multiple disks, if cgpt raises an
error for any of them we print nothing but that error, since we try to
collect and print the disk details at the end. This might also happens
when we use the --all-disks option.

Instead of failing at the first cgpt error, log the errors and print
whatever details we can, but still exit with nonzero status.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>