mkdepthcharge: Don't prepend gpt to kernel cmdline
This reverts commit 9c9715336bec ("mkdepthcharge: Prepend "gpt" to
kernel cmdline"). Using "gpt" in the kernel command line apparently can
cause misdetection when there are remains of a GPT on a device later
reformatted to use MBR [1].

We still want to make "gpt" the default, but making mkdepthcharge do
that makes it harder to avoid adding the parameter with depthchargectl.

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=260040

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>