depthchargectl: Avoid setting argparse nargs to SUPPRESS
Using SUPPRESS as nargs doesn't work on Python 3.6 because of a missing
case in the argparse library, though it was fixed in version 3.7. There
doesn't seem to be a difference between setting nargs to SUPPRESS vs
zero. Set nargs to zero in this case to let the target subcommand run in
Python 3.6 as well.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>