utils: argparse: Make FunctionBindAction support argparse.SUPPRESS dest
Subcommands can set an Argument's 'dest' to argparse.SUPPRESS to avoid
conflicting with their parent's options that they include as a group and
want to extend. This results in '==SUPPRESS==' being added to the parser
result namespace. Instead, do nothing in FunctionBindAction if we pass
argparse.SUPPRESS or None as the 'dest' argument.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>