depthchargectl: build: Demote the uncompressed build skip warning
The "Inputs are too big" warning happens before the "Building image for
board" message, which is a bit confusing when shown as part of a system
upgrade. The warning is also annoying and unneeded, we already show an
error message if we can't build a small enough image. So demote it to an
info message which is only shown with the --verbose argument.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>