depthchargectl: remove: Fix message with missing format call
When calling the remove subcommand on a custom image that is not in the
images dir, an info message would be printed to indicate it's not being
deleted. But the message is missing a format call, so prints the '{}'
placeholder instead of the image path. Add the format call to fix it.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>