depthchargectl: list: Add option to print the count of partitions
Knowing the count of partitions is helpful when implementing system
upgrade hooks. Normally we could use `list -n -o PATH | grep -vc ^$` but
that's ugly. And the obvious way of using `wc -l` to count the parts
doesn't work because print() outputs an empty line when the list is
empty. Add a --count option so that we can do it more easily.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>