utils: argparse: Unbind arguments and groups when calling commands
The Argument and Group objects will bind to the class in which they were
defined if something tries to access the attribute. This happens, for
example, when a subcommand tries to import a global options group from its
parent command by assigning the attribute also to itself.

This results in the argument/group methods not being called because
the objects check the early-bound parent command class for arguments,
instead of the subcommand object that we are working on.

Unbind the existing bound classes from the Arguments and Groups as we
prepare to pass arguments to them, which will make them properly bind to
and work on the subcommand object when we call getattr() later.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>