depthchargectl: config: Query command's config section by default
A previous command updates the commands' config section with the
final values of the config options, like the codename of the detected
board. Make the config subcommand return values from its own config
section by default, instead of the parser-default 'depthcharge-tools',
to let us query the effective values of the config options.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>