depthchargectl: build: Check the initramfs size before building
The default configuration in Debian Unstable arm64 ends up generating an
initramfs of size about 48MiB just by itself. It will definitely not fit
in the 32MiB size limit we have on some arm64 boards no matter how
compress the kernel. This is probably the case for most distros.

Exit early with an error when the initramfs is larger than the image
maximum size, instead of repeatedly rebuilding the image with different
kernel compressions.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>