depthchargectl: write: Avoid "None" in image build error message
When the write subcommand is called without an image or kernel version,
it will try to build an image with a kernel version of None, to use the
build command defaults. If the build command raises an error, it prints
an error message with the value it passed in, which is None. Instead,
don't mention the kernel version if it's None.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>