depthchargectl: Override board argument by copying it
The subcommands override the board argument to for example raise an
error if the board is unknown, or to ignore errors with board argument
processing. But this is done in a hacky way by creating new positional
arguments that we try to disable with nargs=0 and argparse.SUPPRESS.

However these arguments still leak through in edge cases, like when we
have another required positional argument we don't provide, the error
message references our pseudo-disabled argument as a required one.

Now that we can override parent Arguments with the same option strings,
and copy Argument objects to make that easier, do it for the board
argument for all subcommands to get rid of this hack.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>