utils: argparse: Keep inputs in instance dict while calling methods
When preparing Arguments for the Command instances we first bind the
command line arguments to the Argument-bound methods and place them into
the instance dictionary. After all arguments are bound, we call the
Argument-bound methods with the bound arguments in the instance dicts,
and their returns are again put into the instance dict.

The value in the instance dict is removed during the call, so that we
can further call the parent's Argument-bound method with arguments
different than what we were given. However, this makes certain Argument
dependency cycles very hard to implement, where the bound method ends up
returned to the dependents which can't do anything with it because they
don't know which arguments they should pass into it.

Since nothing seems to be calling the parent methods with actually
custom arguments, keep the input bindings in the instance dict while
calling these Argument-bound methods.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>