utils: os: Add a Disks method to get mountpoints of a device
To build images for an offline system we will need to know where its
root and boot partitions are mounted in the running system, so that we
can explore those for the kernel parts.

Add a method to the Disks to get these mountpoints. Exclude results from
/etc/fstab by default because they might not be mounted. Due to bind
mounts and such, a partition can be mounted to multiple places at once,
so return all that we can.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>