depthchargectl: build: Improve mountpoint and root cmdline handling
Having to rederive the root/boot mountpoints from the root argument
every time we want to look into the mounted paths is annoying, and
trying to support multiple mountpoints makes it very ugly.

A previous commit handles mounts that are not mounted from the root of
the source filesystem, which means all mountpoints that we find for a
device usually would have the same content. So choose the shallowest
mountpoint for root and boot, and store them as hidden arguments.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>