depthchargectl: build: Warn if boot partition isn't mounted
If there's a boot partition the kernel parts we need are usually in that
partition, so print a warning if it isn't mounted. The build will either
fail because there are no kernel files found, or should work since they
are specified. In either case, we don't need/want an error for this.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>