depthchargectl: build: Trigger vmlinuz errors before initramfs, dtbs
Trying to build an image for the running system but for a board with a
different CPU architecture means we filter out all installed kernels. In
this case, the build subcommand can print an error about needing an
initramfs for the root cmdline, because the kernel cmdline argument
checks for that.

However, in this case we also don't have a kernel vmlinuz file, which is
a more important error. Try to trigger that error from the initramfs and
dtbs file arguments so that error is printed instead.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>