depthchargectl: build: Prefer custom root configs over subcommand configs
When a custom root is specified we also try to read configs from that
root's /etc/depthcharge-tools, but a config file specified by the
cmdline arguments should override those. The write subcommand calls the
build subcommand with its config object, but those should not override
the custom root configs.

If ConfigParser objects are passed as the config, the parent command's
config argument already builds on top of it, so maybe it's OK to not
override with values from that in this case.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>