depthchargectl: Move root and mountpoint options to parent command
I think it will be cleaner to handle the custom root stuff in the parent
command. If not, we would also need to apply changes onto the write
subcommand as well so that its running system config doesn't override
the build subcommand's custom root configs.

The cyclic dependency between kernel-cmdline and root is still a bit
problematic, solving that is left for a bit later.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>