depthchargectl: Move more of the root arg handling to parent command
The main difference between the parent command's and build subcommand's
root argument is that the latter converts mountpoints to a string that
we can put into the kernel commandline. So keep only that functionality
in the child command, and move the root-from-cmdline fallback into the
parent command.

Unfortunately, we can get into an infinite recursion when neither root
nor kernel-cmdline is specified in the cmdline, as there is a cyclic
dependency: root -> kernel-cmdline -> config -> root-mountpoint -> root.
Break the cyclic dependency at the config where it should be least
problematic.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>