depthchargectl: write: Consider target the root system to work on
When a target mountpoint or a mounted partition is passed to the write
subcommand, consider it the custom root for the write command execution
as well. This is mainly to read configs from that mountpoint so that the
configs we pass into the build and target commands are not about the
running system, but the target system.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>