depthchargectl: build: Avoid calling system root a mountpoint
When the root is a device that doesn't have a UUID and is not listed in
/etc/fstab (e.g. device-mapper stuff), the build subcommand raises an
error because it doesn't have a value to put into the kernel cmdline.
But the error is ugly for the running system because it prints the
mountpoint and calls it a mountpoint. Split the error into two different
messages based on whether we're building for the running system or not.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>