utils: argparse: Improve support for nested Groups
This makes Groups able to somewhat properly wrap Arguments and other
Groups, so that we can add @Group.add on top of an @Argument() chain
which already creates an implicit group for a single method.

Nesting groups is slightly problematic because argparse doesn't really
support them [1]. Most important issue for us is that the inner groups
don't show up in the help message, but that is easily solved by
flattening the arguments into the outermost group.

[1] https://github.com/python/cpython/pull/30098

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>