mkdepthcharge: Implement initramfs support for zimage format
This somewhat manages to hack initramfs support into Depthcharge. More
recent boards have a high enough kernel buffer size that we can put the
initramfs in a place where the kernel won't accidentally overwrite it.
Then we can tell the kernel the precalculated location where it ends up.

The entire body is loaded at the start of the Depthcharge kernel buffer
controlled by its KERNEL_START config, which is by default 0x100000 for
x86 boards. Since Depthcharge jumps directly to that address we can't
put the kernel anywhere except the start (unless we add an intermediate
bootloader). It occupies a range from 1MiB upto 1MiB + bzImage size.

The kernel itself has a dedicated range of memory it will use for
decompression, the decompressor first copies the compressed kernel to
the end of this buffer and then decompresses it from start to end. It
occupies a range from the "preferred address" (usually 0x1000000) of the
kernel upto an "init size", both of which are available for inspection
in the bzImage header.

If we naively add an initramfs to the depthcharge image, it ends up in
the decompression buffer and gets clobbered. But, since we know all the
addresses and offsets, we pad the vmlinuz to the point that anything
appended to it will end up after the decompression buffer.

Then, we can actually append the initramfs to the padded kernel file,
tell the kernel the initramfs' address and size with an "initrdmem"
kernel command line parameter it will use the initramfs.

Instead, this does what I believe Depthcharge and vboot_reference should
be doing: Reclaim the "bootloader" field for the initramfs and set
"ramdisk_image" and "ramdisk_size" values in the Linux boot parameters.
The "bootloader" is unused except on the first few Chromebooks predating
Depthcharge, and even reused this way as a ramdisk for multiboot files.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>