depthchargectl: Fix breaking config cyclic dependency
Commit 464b53d76234 ("depthchargectl: Move more of the root arg
handling to parent command") tries to break up a cyclic dependency on
the config argument. But it inadvertently makes the build subcommand's
board argument use an empty config parser, making it unable to recognize
any board given in the command line argument. This is because there is a
huge cyclic dependency alternating between the parent and child commands:

  _build.py   [293]: root = super().root
  __init__.py [176]: cmdline = self.kernel_cmdline
  _build.py   [343]: self.kernel_cmdline = cmdline = super().kernel_cmdline
  __init__.py [685]: cmdline = self.config.get("kernel-cmdline")
  __init__.py [361]: root = self.root_mountpoint
  __init__.py [221]: if self.root in ("", "None", "none", None):
  _build.py   [293]: root = super().root
  __init__.py [176]: cmdline = self.kernel_cmdline
  _build.py   [381]: if self.initrd is None and self.root is not None:
  _build.py   [178]: self.kernel
  _build.py   [158]: if file_ is None and self.kernel_version is not None:
  _build.py   [93 ]: kernel_arches = self.board.arch.kernel_arches
  _build.py   [131]: board = super().board
  __init__.py [442]: for sectname, section in self.config.parser.items()
  __init__.py [333]: self.config = configparser.ConfigParser()["DEFAULT"]

Eventually we get to how the config argument breaks the dependency,
which makes the board argument raise the error. Instead of breaking the
dependency with an empty config, break it with a "/" root mountpoint.
This makes the config semi-complete for cyclic dependencies and avoids
errors due to empty configs.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>