depthchargectl: target: Print warning for skipped partitions
When the write command tries to use an image too big for any parititon
on the targeted device, it ends up quitting with a "No usable paritions
found" message. This is a bit confusing if we know the device actually
has parititions. Print a warning message about skipping small partitions
so the error makes more sense.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>