depthchargectl: Correctly set images-dir under custom boot mountpoint
When we use a custom root mountpoint with an external /boot partition,
the substituting the root in the default images-dir path results in the
images going under a "boot" directory in the root partition, instead of
to the /boot partition. Interpret paths under /boot as being under the
boot partition as we should.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>