depthchargectl: write: Use chosen root for target call
After building the image, the write subcommand calls the target
subcommand to check which partition it should write the image to, but
this targets based on the current system's bootable disks. Pass the root
argument of the write subcommand to the target subcommand so it targets
the target root's bootable disks.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>