depthchargectl: Break cyclic argument dependencies better
Passing a device description as the root to the write subcommand tries
to target the running system instead of the given root's system. This is
because we break argument dependencies at the root mountpoint with the
current system being a default value.

There are actually two cycles that intersect at the root mountpoint now,
one is roughly root -> kernel_cmdline -> config -> root, and the other
is root -> diskinfo -> root. Break the first one at the kernel_cmdline,
which is apparently the least worst choice. Break the second at the
diskinfo since that's the only other option.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>