mkdepthcharge: Try to fit zimage initrd between kernel and phys addr
The PHYSICAL_START kernel config is used to define the decompression
buffer and is at 10MiB by default. It's marked as an expert config that
shouldn't be changed. Looking at a few AMD64 distro installers, all of
them kept that value and have file sizes around 10MiB.

Which means we almost always have to put the initramfs after preferred
address + init size. This init size is slightly more than uncompressed
vmlinux, which makes the final zimage size extremely large with no
chance to fit in earlier x86 Chromebooks which have a 32MiB kernel
limit.

It looks like if we set the PHYSICAL_START address much higher, we can
put the initramfs before that address. Check for this extremely unlikely
case and try to do so in order to reduce final image size.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>