utils: subprocess: Return Path objects or bytes from compressors
We have been returning CompletedProcess objects from the compression
executables in the subprocess helpers' methods. To cover up for that, we
had wrapped those to provide Path objects in pathlib helpers.

The runner methods in subprocess helpers are meant to be more pythonic
than that. Return Path objects when the destination is a path, or bytes
when it is None. Then remove the wrappers in the pathlib helpers, and
update the callers for the subprocess versions.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>