utils: subprocess: Fix using None for piped runner stdin/stdout
The runners default to using pipes for the process file descriptors
instead of using the Python process' streams, but passing in None as the
stdin/stdout argument overrides this. It shouldn't, explicitly set the
files to subprocess.PIPE when they are none. If we wanted to use the
script's stdin/stdout we would pass sys.stdout etc.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>