depthchargectl: build: Check if we skipped the only compression level
If the inputs are too big for the size limit we skip the uncompressed
build. Raise a size error immediately after this if there is no other
requested/supported compression level. Normally this falls through the
following for loop's else block and raises the same error, but we want
to assume the output file exists there.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>