depthchargectl: build: Don't suggest reducing initramfs if it won't work
After building an image with initramfs too big for the board, we print
an error message which suggests making the initramfs smaller might make
it fit the limit. But this might not be the case for the zimage format,
where the padding we add to make sure initramfs is not overwritten is
itself too big for the size limit. Specifically check if truncating the
initramfs would make the image fit, and only complain about it being big
in that case.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>