depthchargectl: bless: Prioritize partition as the last step
The bless subcommand sets the partition's tries to one, prioritizes it,
and sets/unsets its successful flag, in that order. When writing a new
image to an already successful partition, this means there is a moment
where the untested image is prioritized and marked successful. If the
command is interrupted then and the image fails to boot, depthcharge
won't fallback to a known-good partition because of this.

Prioritize partitions after setting/unsetting their successful flag to
avoid this intermediate maybe-unbootable state.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>