depthchargectl: list: Fix listing bootable disks with missing fstab
The list subcommand can inadvertently fall back to the current system's
bootable disks with a custom root mountpoint if the custom root does not
have the root/boot entries in its /etc/fstab. In this case we should
fallback to the devices mounted to the custom root/boot mountpoints
given as arguments. Fix the bootable disks logic for this and move it to
the list subcommand.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>