mkdepthcharge: Support adding ramdisk load address for FIT images
The earliest 32-bit ARM Chromebooks use U-Boot instead of Depthcharge,
but they use a custom build that implements the same verified boot flow.
Since that flow hooks up into common U-Boot code it supports using an
initramfs in the FIT image, but it requires an explicit load address to
move the initramfs ramdisk image to. Add a mkdepthcharge option to
specify this, and put the given address into the FIT ramdisk subimage
section.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>