utils: Add a method to check for Libreboot firmware ID
Libreboot r20160907 uses Depthcharge for veyron-speedy, but flashing the
firmware on those boards keep the original GBB and therefore HWID. The
Depthcharge it uses is newer than stock and consequently needs a
different board config section. Add utilities to check if a board has
Libreboot-with-Depthcharge based on the firmware ID it replaces.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>