depthchargectl: Remove "gpt" from default kernel cmdline
This reverts commit 01e51de5da27 ("depthchargectl: Add "gpt" to the
default kernel cmdline") and the associated documentation change. The
disk reformatting problem mentioned in that commit is not limited to
some uncommon hardware formatting storage devices improperly. It can
easily be reproduced on any device by writing a Debian Installer iso
image to a GPT formatted disk, meaning the "gpt" parameter will be
a problem much more commonly than the problem it solves.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>