depthchargectl: remove: Keep images if partitions weren't disabled
It's possible that we have the permissions to remove the image file from
the images dir, but not have the permissions to disable partitions on
disks. Do not remove the image if we couldn't fully search the disks and
remove the partitions containing it, as the user would want to re-run
the remove subcommand with the same image.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>