depthchargectl: Implement a systemd kernel-install layout plugin
Copy the 90-loaderentry.install kernel-install plugin from the systemd
repository as of commit 802d9219aa19 ("kernel-install/90-loaderentry: do
not override an existing systemd.machine_id") and modify it to run
depthchargectl remove/write, with some additional checks.

The main complication here is that kernel-install passes our script the
kernel and initramfs files it us wants to use, which I guess might be
different than what depthchargectl would choose on its own. It doesn't
pass in a fdtdir and initramfs files are optional, so we still do some
discovery to specify everything.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>