utils: platform: Ignore errors from crossystem not being installed
Fedora doesn't seem to package crossystem, it doesn't offer to install
it for me when I tried to run it unlike other vboot-utils. Trying to run
it from Python raises a FileNotFoundError. Since reading HWID/FWID works
fine via device-tree on arm64 and ACPI on x86, I'm not sure if running
it is even necessary. But ignore the error keep the call as a last
resort.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>