mkdepthcharge: Increase small post-vmlinuz padding to 256K
If the vmlinuz and initramfs are small enough, we can fit them before
the decompression buffer. But the initramfs still gets corrupted if it
is immediately after the vmlinuz, so we add a small padding between them
vmlinuz in that case.

Another initramfs hack makes it possible to test more distro kernels in
this configuration, which revealed that this padding is still too small.
Fedora, NixOS and Archboot kernels started working at around 0x38000.
Increase it a bit more, to 0x40000. Move its variable to a higher scope
so that it can be used in other places which also need small paddings.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>