mkdepthcharge: Increase init_size to avoid overwriting initramfs
My understanding of the kernel self-decompression so far was that it
uses a fixed memory range at its pref_address (almost always 16MiB) of
init_size size (slightly larger than the uncompressed kernel). More
specifically, that it first copies itself to the end of this buffer,
then decompresses itself to the pref_address.

However, if the kernel was built with KASLR enabled (as is the default)
this decompression address is chosen at random in a way that avoids
overwriting the ramdisk memory area we specify in boot_params. In that
case, the only range we need to avoid is the area overwritten by the
first copy of the compressed vmlinuz.

Apparently, where the vmlinuz copies itself to is calculated using the
init_size value from the boot params at run-time. Increase it enough
that the copy won't overwrite the initramfs in memory, so that it's no
longer necessary to add huge amounts of padding to use an initramfs with
the zimage format.

Suggested-by: Julius Werner <jwerner@chromium.org>
Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>