README.rst: Fix mkdepthcharge invocation
Mkdepthcharge's cmdline option takes multiple arguments until another
option or a double-dash is encountered. So if we specify kernel files
immediately after the cmdline they are wrongly considered part of it,
and mkdepthcharge gives a cryptic "vmlinuz argument is required" error.

This is a pitfall that I also occasionally fall into. It might be better
to make it take only the whole cmdline as a string. For now, just fix
the invocation in the readme file by adding a double-dash.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>