depthchargectl: Support zimage initramfs via init_size hack
Use the init_size hack instead of padding vmlinuz to support initramfs
files for the zimage format, because it can make things small enough to
fit into 32MiB limits.

One problem with this is that it only works when the kernel is built
with KASLR enabled, but that's the default in Linux kconfigs. So provide
a "zimage-initramfs-hack" config option that distributions can set to
"padding" to revert to the previous behaviour.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>