mkdepthcharge: Support multiple initramfs inputs by concatenation
Debian's initramfs files come with the CPU microcode updates already
prepended to it, but Arch Linux's initramfs doesn't. They for example
pass multiple files in grub.cfg which is then responsible for combining
them before passing them to kernel.

Allow mkdepthcharge to take multiple initramfs files as the input,
combine them and process the result as if it was a single initramfs file.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>