depthchargectl: build: Prepend microcode updates to distro initramfs
Arch Linux doesn't itself prepend microcode update images to their
initramfs images, expecting the bootloader to do it. Their files are
/boot/amd-ucode.img and /boot/intel-ucode.img. Check these paths and a
few more that apparently might be microcode update archives. Add them to
the initramfs file that we get from searching the system, but not when
initramfs files are manually specified.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>