depthchargectl: Use /etc/kernel/ and /proc cmdlines by default
Systemd's kernel-install reads /etc/kernel/cmdline for the system's
cmdline. We have used our config.ini for this, but should support that
as well since it looks like a reasonable standard.

If kernel-cmdline is not set in config.ini, defaut to a kernel cmdline
from /etc/kernel/cmdline or /usr/lib/kernel/cmdline, and from
/proc/cmdline if none of those is set. Unset the kernel-cmdline in our
config.ini.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>