mkdepthcharge: Pad for kernel self-decompression in ARMv7 boards
ARMv7 boards suffer from kernel decompression in a similar way to x86
boards, but the start of the decompression buffer is before the address
depthcharge puts our FIT image, and the first copy of the compressed
kernel is just past the end of the decompression result. The hard part
is actually finding the decompressed kernel's size.

Search for common archive magic numbers and try to decompress from
there, as inspired by the extract-vmlinux script from Linux. Then
calculate the lowest address our initramfs can be at, and pad the
vmlinuz according to that. It might be better to add a padding image
entry into the FIT image, but this is fine for now.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>