utils: argparse: Print each element of lists and tuples separately
The depthchargectl remove subcommand returns a list of CrosPartition
objects. Printing this list converts it to a string, but the __repr__
method's output is used for the list items. In this case, we want a list
of partition device paths that CrosPartition returns in its __str__
method.

When a Command program returns a list or a tuple, print each element
separated with new lines. This avoids printing Python-ish text and
leaking implementation details.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>