utils: argparse: Log error messages from returned CommandExit
When a Command returns CommandExit with both message and output, the
caller gets the output and the message is lost. The message here is
intended to be a warning to the caller while we still can provide a
partial output. Log it to the called Command's logger so it's not lost.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>