depthchargectl: Pass mountpoints when calling other subcommands
We might want to avoid adding a root commandline argument to the kernel
cmdline, in which case it's useful to pass --root none. But if we want
to do this for a chroot, we also have to pass --root-mountpoint /mnt so
that it knows where to look for the files. If this is omitted for other
subcommands that e.g. write calls, then they will look for files in the
host root. Pass this and the boot-mountpoint argument to others so they
work correctly.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>