depthchargectl: Include mountpoint arguments in help message
The --root-mountpoint argument is useful for building images for a
chroot without setting a root argument in the kernel command line.
Similarly the --boot-mountpoint argument is useful for when it is not
mounted at the boot subdirectory of root, and not documented in fstab.
Put them in the help message as an acknowledgement that they are
officially supported.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>