README.rst: Mention zimage-initramfs-hack config option
This is a new option that can be "set-init-size" or "pad-vmlinuz",
related to mkdepthcharge options. The reason it's here instead of the
board database is because it's about the vmlinuz, and the reason it's
there at all is that we can't deduce from the file whether set-init-size
will work or not.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>