setup.py: Bump version to 0.6.2
Managed to add a nice initramfs hack for the x86 zimage format thanks to
ChromeOS folks, meaning we are no longer effectively limited to <16MiB
on systems with a 32MiB size limit. I want that for Debian Installer
work, so that prompted more work towards a release. Also solved issues
with ARMv7 Chromebooks since I now have the hardware to test on.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>